streamlit_dashboard/requirements.txt
streamlit==1.40.1
snowflake-connector-python==3.18.0
plotly==5.18.0